pyproject.toml
[project]
name = "meeting-system"
version = "0.1.0"
description = "Multi-role AI architecture review meeting system (MVP)"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn>=0.23.0",
  "langchain-openai>=0.1.0",
  "langchain-core>=0.1.0",
  "dynaconf>=3.2.0",
  "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "httpx>=0.24.0"
]

[dependency-groups]
dev = [
    "ruff>=0.15.0",
    "ty>=0.0.15",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["meeting*"]
exclude = ["config*", "docs*", "examples*", "tests*"]
